requirements.txt
langgraph>=0.2
langchain-openai>=0.1
python-telegram-bot>=21.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.2
python-dotenv>=1.0
pydantic>=2.7
tenacity>=8.3
python-dateutil>=2.9
